Exibição da interface gráfica corrigida (amém) + pequenas correções

+++ Fagocity it/src/fagocity/controller/PlayerController.java
@@ -5,7 +5,7 @@ import fagocity.model.Player;
 public class PlayerController {
 	
 	/* Cria um Player */
-	public static Player createPlayer(double x, double y, ActorID id) {
+	public static Player createPlayer(int x, int y, ActorID id) {
 		Player player = (Player) ActorFactory.createActor(x, y, id);
 		return player;
 	}
